requirements.txt
transformers==4.40.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
accelerate==0.29.3
memory_profiler
fastapi==0.110.3
uvicorn==0.29.0
uuid==1.30
librosa==0.10.2
python-multipart==0.0.9
bitsandbytes==0.41.1
typing==3.7.4.3
ffmpeg-python==0.2.0
pydantic==2.6.4
greenlet==3.0.3
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
cryptography==42.0.7
matplotlib==3.9.0
numpy==1.26.4
openai==1.29.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.1
python-jose==3.3.0
scipy==1.13.0
seaborn==0.13.2
Janome==0.5.0